integrate with ac service
- integrate strafe opti better with current ac service
